digital_transformation/requirements.txt
langchain-community>=0.3.0
langchain-openai>=0.1.1
langchain-core>=0.3.0
langgraph>=0.0.20
pydantic>=2.0.0
tavily-python
duckduckgo-search
wikipedia
python-dotenv
streamlit>=1.32.0 